+++ src/PeriodFactoryTest.php
@@ -36,7 +36,7 @@ final class PeriodFactoryTest extends TestCase
         date_default_timezone_set($this->timezone);
     }
 
-    public function testInstantiationFailsIfTheBoundariesDeclararionAreUnknown(): void
+    public function testInstantiationFailsIfTheboundsDeclararionAreUnknown(): void
     {
         $this->expectException(InvalidTimeRange::class);
         Period::fromDate(new DateTime('2014-01-13'), new DateTime('2014-01-20'), 'foobar');
@@ -380,8 +380,8 @@ final class PeriodFactoryTest extends TestCase
         return [
             'empty string' => ['', 'Y-m-d'],
             'missing separator' => ['[2021-01-02 2021-01-03]', 'Y-m-d'],
-            'missing boundaries' => ['2021-01-02,2021-01-03', 'Y-m-d'],
-            'too many boundaries' => ['[2021-01-02,2021-)01-03]', 'Y-m-d'],
+            'missing bounds' => ['2021-01-02,2021-01-03', 'Y-m-d'],
+            'too many bounds' => ['[2021-01-02,2021-)01-03]', 'Y-m-d'],
             'too many separator' => ['[2021-01-02,2021-,01-03]', 'Y-m-d'],
             'missing dates' => ['[2021-01-02,  ]', 'Y-m-d'],
             'wrong format' => ['[2021-01-02,  ]', 'Ymd'],
@@ -394,18 +394,18 @@ final class PeriodFactoryTest extends TestCase
     public function testCreateNewInstanceFromIsoNotation(
         string $inputFormat,
         string $notation,
-        string $boundaries,
+        string $bounds,
         string $outputFormat,
         string $expected
     ): void {
-        $period = Period::fromIso8601($inputFormat, $notation, $boundaries);
+        $period = Period::fromIso8601($inputFormat, $notation, $bounds);
 
         self::assertSame($expected, $period->toIso8601($outputFormat));
-        self::assertSame($boundaries, $period->boundaries());
+        self::assertSame($bounds, $period->bounds());
     }
 
     /**
-     * @return array<string, array{inputFormat:string, notation:string, boundaries:string, outputFormat:string, expected:string}>
+     * @return array<string, array{inputFormat:string, notation:string, bounds:string, outputFormat:string, expected:string}>
      */
     public function providesValidIso8601Notation(): array
     {
@@ -413,14 +413,14 @@ final class PeriodFactoryTest extends TestCase
             'same input/output format' => [
                 'inputFormat' => 'Y-m-d',
                 'notation' => '2021-03-25/2021-03-26',
-                'boundaries' => '[]',
+                'bounds' => '[]',
                 'outputFormat'=> 'Y-m-d',
                 'expected' => '2021-03-25/2021-03-26',
             ],
             'different input/output format' => [
                 'inputFormat' => 'Y-m-d',
                 'notation' => '2021-03-25/2021-03-26',
-                'boundaries' => '()',
+                'bounds' => '()',
                 'outputFormat'=> 'Y-n-d',
                 'expected' => '2021-3-25/2021-3-26',
             ],
@@ -430,11 +430,11 @@ final class PeriodFactoryTest extends TestCase
     /**
      * @dataProvider provideInvalidIsoNotation
      */
-    public function testFailsToCreateNewInstanceFromIsoNotation(string $notation, string $format, string $boundaries): void
+    public function testFailsToCreateNewInstanceFromIsoNotation(string $notation, string $format, string $bounds): void
     {
         $this->expectException(InvalidTimeRange::class);
 
-        Period::fromIso8601($format, $notation, $boundaries);
+        Period::fromIso8601($format, $notation, $bounds);
     }
 
     /**
@@ -445,7 +445,7 @@ final class PeriodFactoryTest extends TestCase
         return [
             'empty string' => ['', 'Y-m-d', Period::INCLUDE_ALL],
             'missing separator' => ['2021-01-02 2021-01-03', 'Y-m-d', Period::INCLUDE_ALL],
-            'invalid boundaries' => ['2021-01-02/2021-01-03', 'Y-m-d', 'foobar'],
+            'invalid bounds' => ['2021-01-02/2021-01-03', 'Y-m-d', 'foobar'],
             'too many separator' => ['2021-01-02/2021-/01-03', 'Y-m-d', Period::INCLUDE_ALL],
             'missing dates' => ['2021-01-02/', 'Y-m-d', Period::INCLUDE_ALL],
             'wrong format' => ['2021-01-02/2021-01-03', 'Ymd', Period::INCLUDE_ALL],